requirements.txt
# requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart
torch>=1.9.0
torchvision>=0.10.0
pillow>=8.3.1